--- test/example/example_test.txt
@@ -1,12 +1,0 @@
-package example
-
-import "testing"
-
-func TestExample(t *testing.T) {
-	got := 1
-	want := 1
-
-	if got != want {
-		t.Errorf("Got %q, expected %q", got, want)
-	}
-}
