+++ test/samples/format/nodeVar.ygl
