--- test/testData/snippets/testFindTokens.ygl
@@ -1,7 +1,0 @@
-node A {
-
-}
-
-node B {
-
-}
